Renamed GameManager and InstructorGameManager into PilotManager and InstructorManager
