+++ Assets/PilotManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 36d26ea786a07ae419202509f2f2500b
+guid: ddc945a7769b5764eb5c2c2bd2de4e32
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
