+++ Assets/Scenes/PilotGame.unity
@@ -846,7 +846,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 215f9198ef7de44d2b0024ae9a96c0b6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  gameManager: {fileID: 1563405147}
+  pilotManager: {fileID: 1563405149}
 --- !u!1 &1068180337
 GameObject:
   m_ObjectHideFlags: 0
@@ -1339,28 +1339,14 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1563405148}
-  - component: {fileID: 1563405147}
+  - component: {fileID: 1563405149}
   m_Layer: 0
-  m_Name: GameManager
+  m_Name: PilotManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &1563405147
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1563405146}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ddc945a7769b5764eb5c2c2bd2de4e32, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  maxHealth: 1
-  currentHealthAmount: 0
 --- !u!4 &1563405148
 Transform:
   m_ObjectHideFlags: 0
@@ -1376,6 +1362,20 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1563405149
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1563405146}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ddc945a7769b5764eb5c2c2bd2de4e32, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHealth: 3
+  currentHealthAmount: 0
 --- !u!1 &1600300343
 GameObject:
   m_ObjectHideFlags: 0
